Third Commit
it said copy/pate. Now it says copy/paste. That reads better.

yes, a whole commit to correct a typo, but it's easier than modifying a commit. It's actually almost as easy as a save button. If you pull it up you'll see a tiny bit of green and red or yellow there.